+++ src/gospatial/app/logging.go
@@ -2,7 +2,7 @@ package app
 
 import (
 	"io"
-	"io/ioutil"
+	// "io/ioutil"
 	"log"
 	"os"
 )
@@ -39,17 +39,8 @@ func DebugMode() {
 	Error = log.New(DebugModeLogFile, "[FIND] ERROR | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
 }
 
-func TestMode_old() {
-	AppMode = "testing"
-	Trace = log.New(ioutil.Discard, "[FIND] TRACE | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
-	Info = log.New(ioutil.Discard, "[FIND] INFO  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
-	Debug = log.New(ioutil.Discard, "[FIND] DEBUG | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
-	Warning = log.New(ioutil.Discard, "[FIND] WARN  | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
-	Error = log.New(ioutil.Discard, "[FIND] ERROR | ", log.LUTC|log.Ldate|log.Ltime|log.Lshortfile|log.Lmicroseconds)
-}
-
 func TestMode() {
-	AppMode = "testing"
+	AppMode = "test"
 	DebugModeLogFile, err := os.OpenFile("test.log", os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0666)
 	if err != nil {
 		Error.Fatal("Error opening file: %v", err)