+++ src/gospatial/main/gospatial_backup/main.go
@@ -53,7 +53,7 @@ func main() {
 		if err != nil {
 			app.Error.Fatal(err)
 		}
-		app.Info.Println(string(file))
+		// app.Info.Println(string(file))
 		// unmarshal data
 		var data DumpedDatabase
 		data.Apikeys = make(map[string]app.Customer)
@@ -63,13 +63,15 @@ func main() {
 			app.Error.Fatal(err)
 		}
 		// load api keus
-		for k := range data.Apikeys {
-			app.DB.InsertCustomer(data.Apikeys[k])
-		}
+		// for k := range data.Apikeys {
+		// 	app.DB.InsertCustomer(data.Apikeys[k])
+		// }
+		app.DB.InsertCustomers(data.Apikeys)
 		// load lauers
-		for k := range data.Layers {
-			app.DB.InsertLayer(k, data.Layers[k])
-		}
+		// for k := range data.Layers {
+		// 	app.DB.InsertLayer(k, data.Layers[k])
+		// }
+		app.DB.InsertLayers(data.Layers)
 	}
 
 	os.Exit(0)