+++ CHANGELOG.md
@@ -2,7 +2,6 @@
  - ping route
  - TLS (wss:// && https://) 
  - pprof
- - On db_init put customers into memory
  - Testing with Caddy --> https://github.com/mholt/caddy
  
 
@@ -10,6 +9,8 @@
 
 ## Unreleased 1.6.4
 
+### Added
+ - On db_init put customers into memory